Update AlphaBeta : condition is now >=, before it was > only.

+++ OthelloIA_G3/TreeNode.cs
@@ -23,10 +23,11 @@ namespace OthelloIA_G3
         }
 
         /// <summary>
-        /// La valeur retournée indique la différence de coups possible entre les joueurs. Si positif : noir à l'avantage.
+        /// V1 : La valeur retournée indique la différence de coups possible entre les joueurs. Si positif : noir à l'avantage.
+        /// V2 : Division Noir/Blanc : si val > 1 alors black a l'avantage.
         /// </summary>
         /// <returns></returns>
-        public int Eval()
+        public double Eval()
         {
             int possibilityWhite = Ops(true).Count;
             int possibilityBlack = Ops(false).Count;
